feat: make client event is received in server
- Remove DOM notion
