+++ lib/js/src/render_html.js
@@ -26,7 +26,7 @@
 function renderHtml(root, widget, replace = false) {
     const el = document.createElement(widget.tag);
     el.id = widget.id;
-    el.style = widget.style
+    el.style = widget.style;
     el.innerHTML = widget.content;
 
     for (const attr of widget.attributes) {
@@ -38,8 +38,16 @@ function renderHtml(root, widget, replace = false) {
             console.log({
                 event: event.name,
                 target: widget.tag,
-                id: widget.id
+                id: widget.id,
+                action: widget.tag + "-" + event.name + "-" + widget.id,
             });
+
+            socket.send(
+                JSON.stringify({
+                    action: widget.tag + "-" + event.name + "-" + widget.id,
+                    data: {},
+                })
+            );
         });
     }
