+++ lib/utils.go
@@ -1,30 +1,11 @@
 package lib
 
 import (
-	"fmt"
 	"math/rand"
 	"time"
 )
 
-func (w *Widget) generateId(length int) string {
-	const charset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-	seededRand := rand.New(rand.NewSource(time.Now().UnixNano()))
-	b := make([]byte, length)
-	for i := range b {
-		b[i] = charset[seededRand.Intn(len(charset))]
-	}
-	return string(b)
-}
-
-func (w *Widget) renderOpenTag() string {
-	return fmt.Sprintf("<%s id=\"%s\" ", w.GetData("tag").(string), w.id)
-}
-
-func (w *Widget) renderCloseTag() string {
-	return fmt.Sprintf("</%s>", w.GetData("tag").(string))
-}
-
-func generateRandomId(length int) string {
+func generateId(length int) string {
 	const charset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
 	seededRand := rand.New(rand.NewSource(time.Now().UnixNano()))
 	b := make([]byte, length)
