+++ lib/widget.go
@@ -5,9 +5,10 @@ import (
 	"strings"
 )
 
+var widgetPool = make([]string, 0)
+
 type WidgetTree map[int]*Widget
 type WidgetRender func(obj *RenderHtmlPayload)
-type WidgetEvents map[string]EventHandler
 
 type Widget struct {
 	id       string
@@ -15,27 +16,32 @@ type Widget struct {
 	render   WidgetRender
 	data     WidgetData
 	children WidgetTree
-	events   WidgetEvents
+	events   []WidgetEvent
 	style    WidgetStyle
 
 	index  int
 	parent *Widget
 }
 
+func (w *Widget) GetId() string {
+	return w.id
+}
+
 func newWidget() *Widget {
+	id := generateId(10)
+
+	widgetPool = append(widgetPool, id)
 	return &Widget{
+		id:       id,
 		data:     make(WidgetData),
 		children: make(WidgetTree),
-		events:   make(WidgetEvents),
+		events:   make([]WidgetEvent, 0),
 		style:    NewWidgetStyle(),
 	}
 }
 
 func (w *Widget) Delete() {
-	// parent := w.parent
-	dom.RemoveWidget(w.id)
 	w.parent.RemoveChild(w.index)
-	// parent.emitContentUpdate()
 }
 
 func (w *Widget) Dump(identLevel int) {
@@ -61,7 +67,7 @@ func (w *Widget) AddChild(children ...*Widget) error {
 		children[i].index = index
 	}
 
-	dom.Register(w)
+	// dom.Register(w)
 	return nil
 }
 
@@ -111,27 +117,126 @@ func (w *Widget) Render() *RenderHtmlPayload {
 	return obj
 }
 
-// TODO: move to widgetAttribute.go
+// ATTRIBUTE SECTION //
+
 type WidgetAttribute struct {
 	Name  string `json:"name"`
 	Value string `json:"value"`
 }
 
-// TODO: move to widgetEvent.go
+// EVENT SECTION //
 type WidgetEvent struct {
 	Name string `json:"name"`
 }
 
 func (w *Widget) Events() []WidgetEvent {
-	events := make([]WidgetEvent, 0, len(w.events))
-	for name, _ := range w.events {
-		events = append(events, WidgetEvent{
-			Name: name,
-		})
-	}
-	return events
+	return w.events
 }
 
-func (w *Widget) SetEvent(name string) {
-	w.events[name] = nil
+func (w *Widget) SetEvent(name string, handler EventHandler) {
+	w.events = append(w.events, WidgetEvent{name})
+
+	// tag-event-id
+	key := fmt.Sprintf("%s-%s-%s", w.GetData("tag").(string), name, w.id)
+	registerEvent(key, handler)
+}
+
+// DATA SECTION //
+
+type WidgetData map[string]interface{}
+
+func (w *Widget) SetData(key string, value interface{}) {
+	w.data[key] = value
+	// w.emitContentUpdate()
+}
+
+func (w *Widget) GetData(key string) interface{} {
+	return w.data[key]
+}
+
+func (w *Widget) HasData(key string) bool {
+	_, ok := w.data[key]
+	return ok
+}
+
+func (w *Widget) DeleteData(key string) {
+	delete(w.data, key)
+	// w.emitContentUpdate()
+}
+
+func (w *Widget) ClearData() {
+	w.data = make(WidgetData)
+	// w.emitContentUpdate()
+}
+
+// KIND SECTION //
+
+type WidgetKind int
+
+const (
+	WidgetElement = iota
+	WidgetLabel
+	WidgetButton
+	WidgetGrid
+)
+
+func (wk *WidgetKind) String(w *Widget) string {
+	switch *wk {
+	case WidgetElement:
+		return fmt.Sprintf("WidgetElement(%s)", w.GetData("tag").(string))
+	case WidgetLabel:
+		return "WidgetLabel"
+	case WidgetButton:
+		return "WidgetButton"
+	case WidgetGrid:
+		return fmt.Sprintf("WidgetGrid(%d, %d)", w.GetData("rows").(int), w.GetData("cols").(int))
+	}
+	return "WidgetKind"
+}
+
+// STYLE SECTION //
+type WidgetStyle struct {
+	data map[string]interface{}
+}
+
+func NewWidgetStyle() WidgetStyle {
+	return WidgetStyle{
+		data: make(map[string]interface{}),
+	}
+}
+
+func (w *WidgetStyle) Set(key string, value interface{}) {
+	w.data[key] = value
+}
+
+func (w *WidgetStyle) Get(key string) interface{} {
+	return w.data[key]
+}
+
+func (w *WidgetStyle) String() string {
+	style := ""
+	for key, value := range w.data {
+		style += fmt.Sprintf("%s: %s; ", key, value)
+	}
+
+	return style
+}
+
+func (w *Widget) SetStyle(key string, value interface{}) {
+	w.style.Set(key, value)
+	// w.emitContentUpdate()
+}
+
+func (w *Widget) GetStyle(key string) interface{} {
+	return w.style.Get(key)
+}
+
+func (w *Widget) HasStyle(key string) bool {
+	_, ok := w.style.data[key]
+	return ok
+}
+
+func (w *Widget) DeleteStyle(key string) {
+	delete(w.style.data, key)
+	// w.emitContentUpdate()
 }
